Adapt "Visit flagged resources" test to include proposal_notification as factory
Note that the proposal notifications links were pointing to the
`proposal_notifications#show` action, which moderators can't visit
because only the author of the notification can visit that URL. So we're
changing the URL; now it points to the proposal containing the
notification.

Also note that that notification don't have flags; if we moderate them,
they become immediately hidden. So, in the test, we don't visit the
"Pending" tab but the "All" tab instead. Visiting the page of a
moderated proposal notification doesn't show the contents of that
proposal notification, since that notification is already hidden.