Simplify moderation tests that hide content
We can simply refresh the page after hiding a certain content and check
that the content is no longer present.

When blocking the author, we're checking the author has been blocked
instead of checking the current path. This way we can remove all the
`index_path` logic in the tests.